requirements.txt
Django~=4.0.3
Pillow
psycopg2-binary
djangorestframework
drf-yasg
django-filter
python-decouple
djangorestframework-simplejwt
django-crispy-forms
